backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
alembic==1.12.1
python-dotenv==1.0.0
pyjwt==2.8.0
apscheduler==3.10.4
# Chatbot & RAG dependencies - Updated compatible versions
openai>=1.6.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.15
faiss-cpu>=1.7.4
tiktoken>=0.5.0 